requirements.txt
elasticsearch==8.17.0
python-dotenv==1.0.1
requests==2.32.3
pandas==2.2.3
numpy==1.26.4
faker==24.9.0
pytest==8.2.0

